[rename] account -> currentAccount
