--- apps/fe/src/store/features/account/accountSlice.ts
@@ -1,24 +1,0 @@
-import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
-import type { Account } from "./type";
-
-const initialState: Account = {
-  username: "",
-  email: "",
-  password: "",
-  message: "",
-};
-
-export const AccountSlice = createSlice({
-  name: "account",
-  initialState,
-  reducers: {
-    updateAccountInfo: (state, action: PayloadAction<Partial<Account>>) => {
-      return { ...state, ...action.payload };
-    },
-    clearAccountInfo: () => initialState,
-  },
-});
-
-export const { updateAccountInfo, clearAccountInfo } = AccountSlice.actions;
-
-export default AccountSlice.reducer;
