+++ apps/fe/src/store/features/account/currentAccountSlice.ts
@@ -1,0 +1,24 @@
+import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
+import type { Account } from "./type";
+
+const initialState: Account = {
+  username: "",
+  email: "",
+  password: "",
+  message: "",
+};
+
+export const CurrentAccountSlice = createSlice({
+  name: "currentAccount",
+  initialState,
+  reducers: {
+    updateCurrentAccountInfo: (state, action: PayloadAction<Partial<Account>>) => {
+      return { ...state, ...action.payload };
+    },
+    clearCurrentAccountInfo: () => initialState,
+  },
+});
+
+export const { updateCurrentAccountInfo, clearCurrentAccountInfo } = CurrentAccountSlice.actions;
+
+export default CurrentAccountSlice.reducer;
