+++ apps/fe/src/store/store.ts
@@ -3,7 +3,7 @@
 import { combineReducers, configureStore } from "@reduxjs/toolkit";
 import { persistReducer, persistStore } from "redux-persist";
 import storage from "redux-persist/lib/storage";
-import accountReducer from "./features/account/accountSlice";
+import currentAccountReducer from "./features/account/currentAccountSlice";
 import accountsReducer from "./features/account/accountsSlice";
 import postReducer from "./features/post/postSlice";
 
@@ -14,7 +14,7 @@ const persistConfig = {
 
 const rootReducer = combineReducers({
   // 新規作成したReducerはここに追加していく
-  account: accountReducer,
+  currentAccount: currentAccountReducer,
   accounts: accountsReducer,
   post: postReducer,
 });
